Updated domain model diagram, and sequence-3

+++ app/src/main/java/com/example/fitguide/domain/model/User.kt
@@ -13,7 +13,7 @@ data class `User`(
     val lastName: String? = null,
     val email: String? = null,
     val gender : String? = null,
-    val age: Int? = null, // date of birth in milliseconds since epoch
+    val age: Long? = null, // date of birth in milliseconds since epoch
     val height: Float? = null,   // height in cm
     val weight: Float? = null,   // weight in kg
     val coachId: Long? = null, // If the user has a coach.
